version 0.3
I noticed a weakness in the vertex due to the larger 2020 extrusion. I
modified the code to move the "wings" further "down". Honestly hart to
describe but you can see the difference in photos of the v0.2 top and
motor.stl files.

Also added the code and stl for the tnut tool.

As before, these are still untested and unsupported :)